Fixed button error and refactored
