+++ app/src/main/java/com/iamtakuu/asap_timesheet_app/data/tasks/TaskCreatedEvent.kt
@@ -1,4 +1,4 @@
-package com.iamtakuu.asap_timesheet_app.data.login
+package com.iamtakuu.asap_timesheet_app.data.tasks
 
 sealed class TaskCreatedEvent {
     data class StartDateChanged(val startDate:String): TaskCreatedEvent()
