+++ app/src/main/res/values/strings.xml
@@ -18,6 +18,7 @@
     <string name="login">Login</string>
 
     <string name="task_details">Task Details</string>
+    <string name="add_image">Add Image</string>
     <string name="start_date">Start Date</string>
     <string name="end_date">End Date</string>
     <string name="time">Time</string>
